+++ CHANGES.rst
@@ -11,6 +11,7 @@ Release date to be decided.
 
 - async support
 - Twisted support
+- concurrent future support
 - pipelines accept crawler on start/stop/process calls
 - downloader middleware also process exception